updating GWAS insert method

+++ src/edu/mcw/rgd/dao/impl/GWASCatalogDAO.java
@@ -31,14 +31,14 @@ public class GWASCatalogDAO extends AbstractDAO {
         BatchSqlUpdate su = new BatchSqlUpdate(this.getDataSource(),"insert into gwas_catalog (gwas_id, pmid, disease_trait, init_sample_size, replicate_sample_size, region, chromosome, pos, report_genes, mapped_genes, snp_risk_allele, " +
                 "snps, cur_snp_id, context, risk_allele_freq, p_value, p_value_mlog, snp_passing_qc, mapped_trait, efo_ids, study_acc) values (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)",
                 new int[]{Types.INTEGER,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,
-                        Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.INTEGER,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR});
+                        Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.FLOAT,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR,Types.VARCHAR});
         su.compile();
         for (GWASCatalog gc : toBeInserted){
             int gwasId = this.getNextKeyFromSequence("GWAS_CAT_SEQ");
             gc.setGwasId(gwasId);
 
             su.update(gc.getGwasId(),gc.getPmid(),gc.getDiseaseTrait(),gc.getInitialSample(), gc.getReplicateSample(),gc.getRegion(),gc.getChr(),gc.getPos(),
-            gc.getReportedGenes(),gc.getMappedGene(),gc.getStrongSnpRiskallele(),gc.getSnps(),gc.getCurSnpId(),gc.getContext(),gc.getRiskAlleleFreq(),gc.getpValStr(),gc.getpValMlog(),
+            gc.getReportedGenes(),gc.getMappedGene(),gc.getStrongSnpRiskallele(),gc.getSnps(),gc.getCurSnpId(),gc.getContext(),gc.getRiskAlleleFreq(),gc.getpValStr(),gc.getpValMlog().doubleValue(),
             gc.getSnpPassQc(),gc.getMapTrait(),gc.getEfoId(),gc.getStudyAcc());
         }
         return executeBatch(su);
